% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettab.R
\name{print.osninfo}
\alias{print.osninfo}
\title{printer for osninfo}
\usage{
\method{print}{osninfo}(x, ...)
}
\arguments{
\item{x}{output of `process_tab`}

\item{\dots}{not used}
}
\description{
printer for osninfo
}
