% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettab.R
\name{probe}
\alias{probe}
\title{blow up some info for a selected component of OSN info}
\usage{
probe(osni, node)
}
\arguments{
\item{osni}{instance of "osninfo", as created with process_tab}

\item{node}{one of the components of osni$split}
}
\description{
blow up some info for a selected component of OSN info
}
\examples{
x = gettab()
y = process_tab(x)
probe(y, "AnnotationHub") |> head(20)
}
