% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettab.R
\name{process_tab}
\alias{process_tab}
\title{process table for hierarchy exploration}
\usage{
process_tab(tab)
}
\arguments{
\item{tab}{output of `gettab`}
}
\description{
process table for hierarchy exploration
}
